Moved helper functions to `meta_rules`
