+++ src/meta_rules/node.rs
@@ -2,13 +2,15 @@ use range::Range;
 use std::rc::Rc;
 use std::cell::Cell;
 
-use {
+use super::{
     ret_err,
     update,
+    ParseResult,
+};
+use {
     DebugId,
     MetaData,
     ParseError,
-    ParseResult,
     Rule,
     Tokenizer,
     TokenizerState,
@@ -89,7 +91,8 @@ impl Node {
 #[cfg(test)]
 mod tests {
     use all::*;
-    use meta_rules::{ Node, Number, Optional, Sequence, Whitespace };
+    use meta_rules::{ update_refs, Node, Number, Optional, Sequence,
+        Whitespace };
     use std::rc::Rc;
     use std::cell::Cell;
