+++ src/meta_rules/sequence.rs
@@ -1,11 +1,13 @@
 use range::Range;
 use std::rc::Rc;
 
-use {
+use super::{
     ret_err,
     update,
-    DebugId,
     ParseResult,
+};
+use {
+    DebugId,
     Rule,
     Tokenizer,
     TokenizerState,
