+++ src/meta_rules/until_any_or_whitespace.rs
@@ -1,11 +1,13 @@
 use read_token;
 use std::rc::Rc;
 
+use super::{
+    ParseResult,
+};
 use {
     DebugId,
     MetaData,
     ParseError,
-    ParseResult,
     Tokenizer,
     TokenizerState,
 };
